requirements.txt
pandas
scikit-learn
mlflow
streamlit
joblib
pyyaml